Extract into function
